// frontend/src/types/generated/GetAdminProductsQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AcidityType } from "./AcidityType";
import type { BodyType } from "./BodyType";
import type { EffervescenceType } from "./EffervescenceType";
import type { IncludeDeleted } from "./IncludeDeleted";
import type { MeadType } from "./MeadType";
import type { SweetnessType } from "./SweetnessType";
import type { TanninsType } from "./TanninsType";
import type { TurbidityType } from "./TurbidityType";

export type GetAdminProductsQuery = { include_deleted?: IncludeDeleted, order_by?: string, order_direction?: string, in_stock?: boolean, product_type?: MeadType, sweetness?: SweetnessType, turbidity?: TurbidityType, effervescence?: EffervescenceType, acidity?: AcidityType, tannins?: TanninsType, body?: BodyType, page?: number, per_page?: number, };