// frontend/src/types/generated/Order.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { OrderStatus } from "./OrderStatus";

export type Order = { order_id: string, status: OrderStatus, currency: string, total_amount_cents: number, stripe_session_id: string | null, stripe_payment_intent_id: string | null, customer_email: string | null, language: string, created_at: string, updated_at: string, };